Render Blog Post Thumbnails on index page and removes other content
