+++ src/data/mocks.js
@@ -1,0 +1,9 @@
+export const mockBlogThumbnails = [
+  {
+    title: 'Blog Post Thumbnail Title',
+    date: '2023-01-01',
+    summary:
+      'This is a sample blog post thumbnail summary. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Donec a purus malesuada, accumsan velit sit amet, feugiat leo. Integer id accumsan erat. Ut feugiat suscipit lectus ac rutrum. Sed vitae venenatis augue. Sed gravida dolor nunc, non varius turpis rhoncus in. Sed felis nisl, ultricies vitae faucibus facilisis, euismod nec dolor.',
+    url: '/',
+  },
+]
