+++ src/pages/component-library.js
@@ -1,4 +1,4 @@
-import { mockBlogThumbnails } from '../data'
+import { mockBlogThumbnails } from '../data/mocks'
 import BackToHomeButton from '../components/back-to-home-button'
 import BlogPostThumbnails from '../components/blog-post-thumbnails'
